requirements.txt
streamlit
sentence-transformers
scikit-learn
numpy
requests
python-dotenv
